解决推荐页面打开UI卡顿，延迟加载

+++ app/src/main/java/com/example/administrator/searchpicturetool/model/bean/NewRecommendContent.java
@@ -15,7 +15,7 @@ public class NewRecommendContent extends BmobObject{
     private String content;
     private boolean justType;
 
-    public double getType() {
+    public float getType() {
         return type;
     }
